Adding delete booking

+++ app/src/main/java/com/example/your_space/repository/AppRepository.kt
@@ -1,6 +1,8 @@
 package com.example.your_space.repository
 
 import android.content.Context
+import android.util.Log
+import android.widget.Toast
 import androidx.lifecycle.LiveData
 import androidx.lifecycle.map
 import com.example.your_space.database.AppDao
@@ -8,12 +10,19 @@ import com.example.your_space.database.AppDatabase
 import com.example.your_space.database.bookingToDomainModel
 import com.example.your_space.database.spaceToDomainModel
 import com.example.your_space.network.Network
+import com.example.your_space.network.Network.NetworkServices
 import com.example.your_space.network.networkdatamodel.bookingProertyModelToDatabaseModel
 import com.example.your_space.network.networkdatamodel.propertyModelToDatabaseModel
 import com.example.your_space.ui.booking.BookItem
 import com.example.your_space.ui.ourspaces.SpaceItem
+import kotlinx.coroutines.CoroutineScope
 import kotlinx.coroutines.Dispatchers
+import kotlinx.coroutines.launch
 import kotlinx.coroutines.withContext
+import okhttp3.ResponseBody
+import retrofit2.Call
+import retrofit2.Callback
+import retrofit2.Response
 
 class AppRepository private constructor(private val database: AppDao) {
 
@@ -70,8 +79,21 @@ class AppRepository private constructor(private val database: AppDao) {
     }
 
     suspend fun deleteBookingWithId(bookItem: BookItem) {
-        withContext(Dispatchers.IO) {
-            database.deleteBooking(bookItem.bookId)
+        CoroutineScope(Dispatchers.IO).launch {
+
+            // Do the DELETE request and get response
+            val response = NetworkServices.cancelBooking(bookItem.bookId.toInt())
+            withContext(Dispatchers.IO) {
+                if (response.isSuccessful) {
+                    database.deleteBooking(bookItem.bookId)
+                    refreshBookings()
+
+                } else {
+
+                    Log.e("RETROFIT_ERROR", response.code().toString())
+
+                }
+            }
         }
     }
